Finished video 118

+++ mean-course/backend/routes/posts.js
@@ -59,8 +59,12 @@ router.put('/:id', checkAuth, multer({storage: storage}).single('image'), (req, 
         content: req.body.content,
         imagePath: imagePath
     });
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-        res.status(200).json({message: 'Update successful'});
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if (result.modifiedCount > 0) {
+            res.status(200).json({message: 'Update successful'});
+        } else {
+            res.status(401).json({message: 'Not authorized'});
+        }
     });
 });
 
@@ -96,10 +100,13 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({message: 'Post deleted!'});
-    })
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+        if (result.deletedCount > 0) {
+            res.status(200).json({message: 'Deletion Successful'});
+        } else {
+            res.status(401).json({message: 'Not authorized'});
+        }
+    });
 });
 
 module.exports = router;
